requirements.txt
langchain
langchain-openai
langchain-chroma
langchain-community
python-dotenv
requests
beautifulsoup4
PyPDF2
chromadb
